Update Modrinth Minotaur
I know @PotatoPresident said he wanted to switch to mc-publish at one point or another, but the switch away from 1.x is a rather urgent one considering dependencies in 1.x are not and have never been supported. As it stands right now, your Minotaur configuration is resulting in dependencies being listed incorrectly. I've gone through and fixed the dependencies on the versions already on the site, but please do not make a new release without first merging this or switching to mc-publish.